move related code and tests to audio directory
