+++ main.py
@@ -8,7 +8,7 @@ import sys
 from pathlib import Path
 from typing import List
 
-from metadata import adjust_metadata
+from audio.metadata import adjust_metadata
 from gui.main_window import show_main_window
 
 
@@ -16,7 +16,6 @@ def get_paths(path: str) -> List[Path]:
     """
     Returns a list of paths from the given arguments.
     """
-
     if not os.path.exists(path):
         print(f"Given path does not exist: {path}")
         return []
@@ -38,14 +37,15 @@ if __name__ == "__main__":
     if len(sys.argv) == 1:
         show_main_window()
 
-    elif len(sys.argv) != 2:
+    elif len(sys.argv) == 2:
+        files = get_paths(sys.argv[1])
+        for file in files:
+            print("processing file: ", file)
+            adjust_metadata(file, comments="", genre="House")
+
+    else:
         print(
             "You need to specify a specific file " +
             "or a directory containing mp3s!"
         )
         sys.exit(1)
-
-        files = get_paths(sys.argv[1])
-        for file in files:
-            print("processing file: ", file)
-            adjust_metadata(file, comments="", genre="House")
